Squashed commit of the following:
commit 5007132ca01cba3aa8686dbbb753bc4076458c49
Merge: 5447cb0 7de896a
Author: shchoon <87121448+shchoon@users.noreply.github.com>
Date:   Wed Jan 8 15:01:27 2025 +0900

    Merge pull request #28 from booksaetong/feat/modal

    [feat/modal] 상세 모달 isbn 수정

commit 7de896a880ba957f633d84557b0828d1628985d9
Author: shchoon <acckr0604@gmail.com>
Date:   Wed Jan 8 14:55:10 2025 +0900

    Modify: 상세 모달 isbn 수정

commit 5447cb04b71c7ffaf0bf612131bdbcce77aefcaa
Merge: bcc9237 b5f65cd
Author: KimYuHyun <103613612+SeanKim05@users.noreply.github.com>
Date:   Wed Jan 8 12:21:20 2025 +0900

    Merge pull request #27 from booksaetong/feat/bookshelf

    [Feat/bookshelf] 서재페이지 상세모달 연동

commit b5f65cdcfe17f33694f57dd735073c10a1f52538
Author: SeanKim05 <dbgus20609@gmail.com>
Date:   Wed Jan 8 11:31:55 2025 +0900

    Fix: 미사용 파일 정리

commit af842fa5165b7d546fa4047a511b33d701de47e7
Author: SeanKim05 <dbgus20609@gmail.com>
Date:   Wed Jan 8 11:23:21 2025 +0900

    Fix: propagation 제거

commit d5586848c488076155515af6dc42f44d8e7f5844
Author: SeanKim05 <dbgus20609@gmail.com>
Date:   Wed Jan 8 11:10:14 2025 +0900

    Fix: modal 누락된 상태 추가

commit d8aff780e2a8b14df4e7c3edbd713c374fbad8d0
Author: SeanKim05 <dbgus20609@gmail.com>
Date:   Wed Jan 8 10:28:10 2025 +0900

    Feat: 상세 모달 연동작업 완료

commit 468aee30bb50e5914ffc30624136b417bc9d8fac
Author: SeanKim05 <dbgus20609@gmail.com>
Date:   Wed Jan 8 10:18:46 2025 +0900

    Feat: 상세모달 연동 작업
